L68 Connecting All Todo App Features to H2 Database

+++ src/main/java/com/miltontest/springboot/miltonfirstwebapp/todo/ToDoRepository.java
@@ -4,6 +4,6 @@ import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ToDoRepository extends JpaRepository<ToDo, Long>{
+public interface ToDoRepository extends JpaRepository<ToDo, Integer>{
   public List<ToDo> findByUsername(String username);
 }
